ref(minidump): Move minidump logic to its own module
